requirements.txt
#requirements.txt
# Dependencias del proyecto
python-chess
PyQt5
matplotlib
seaborn
